+++ CHANGELOG.rst
@@ -25,6 +25,8 @@ API is now at version **1.17**. See `API changelog`_.
 - Fix requests output when running with make serve (fixes #1242)
 - Fix pagination on permissions endpoint (fixes #1157)
 - Fix pagination when max fetch storage is reached (fixes #1266)
+- Fix schema validation when internal fields like ``id`` or ``last_modified`` are
+  marked as required (fixes #1244)
 
 **Internal changes**